multipart-attachments tested and working

+++ library/utility/multipart-attachments.js
@@ -7,42 +7,20 @@ const crypto = require('crypto')
 const codec = require('../models/codec')
 const xbytes = require('xbytes')
 const finished = require('on-finished')
-const os = require('os')
 const fs = require('fs')
-const path = require('path')
 
-function getTempFilePath () {
-  const rand = crypto.randomBytes(20).toString('hex')
-  return path.join(os.tmpdir(), `pigeon-optics-${rand}`)
-}
-
+const memStore = multer.memoryStorage()
+const diskStore = multer.diskStorage({})
 const uploadEngine = {
   _handleFile: async function (req, file, cb) {
     if (file.fieldname === 'body') {
-      const chunks = []
-      for await (const chunk of file.stream) chunks.push(chunk)
-      if (file.mimetype === 'application/cbor') {
-        req.body = codec.cbor.decode(Buffer.concat(chunks))
-      } else if (file.mimetype === 'application/json') {
-        req.body = codec.json.decode(Buffer.concat(chunks).toString('utf-8'))
-      } else {
-        throw createHttpError.BadRequest('Unsupported mime type on body file')
-      }
-      cb()
+      return memStore._handleFile(req, file, cb)
     } else if (file.fieldname === 'attachment') {
-      const path = getTempFilePath()
-      const writeStream = file.stream.pipe(fs.createWriteStream(path))
-      writeStream.on('finished', () => {
-        const hash = fs.createReadStream(path).pipe(crypto.createHash('sha256'))
-        hash.setEncoding('hex')
-        hash.on('data', hex => {
-          cb(null, { path, size: writeStream.bytesWritten, hash: hex })
-        })
-      })
+      return diskStore._handleFile(req, file, cb)
     }
   },
   _removeFile: function (req, file, cb) {
-    fs.unlink(file.path, cb)
+    if (file.path) fs.unlink(file.path, cb)
   }
 }
 
@@ -54,29 +32,51 @@ const multerInstance = multer({
   { name: 'attachment' }
 ])
 
-const postProcess = function (req, res, next) {
+const postProcess = async function (req, res, next) {
   req.attachedFilesByHash = {}
   req.attachedFilesByName = {}
+  const tasks = []
 
   if (req.files) {
-    for (const file of req.files) {
-      if (file.hash) req.attachedFilesByHash[file.hash] = file
-      if (file.originalname) req.attachedFilesByName[file.originalname] = file
+    if (Array.isArray(req.files.attachment)) {
+      for (const file of req.files.attachment) {
+        if (file.originalname) req.attachedFilesByName[file.originalname] = file
+
+        tasks.push(new Promise((resolve, reject) => {
+          const hash = fs.createReadStream(file.path).pipe(crypto.createHash('sha256'))
+          hash.setEncoding('hex')
+          hash.on('data', hex => {
+            file.hash = hex
+            req.attachedFilesByHash[hex] = file
+            resolve()
+          })
+          hash.on('error', reject)
+        }))
+      }
+
+      // clean up after request finishes
+      finished(res, async () => {
+        for (const file of req.files.attachment) {
+          try { await fs.promises.unlink(file.path) } catch (err) {
+            // nothing
+          }
+        }
+      })
+    }
+
+    if (req.files.body && req.files.body[0]) {
+      const bodyInfo = req.files.body[0]
+      if (bodyInfo.mimetype === 'application/json') {
+        Object.assign(req.body, codec.json.decode(bodyInfo.buffer.toString('utf-8')))
+      } else if (bodyInfo.mimetype === 'application/cbor') {
+        Object.assign(req.body, codec.cbor.decode(bodyInfo.buffer))
+      } else {
+        return next(new createHttpError.BadRequest('Unsupported body mime type'))
+      }
     }
   }
 
-  if (req.files && req.files.length) {
-    // clean up after request finishes
-    finished(res, async () => {
-      for (const file of req.files) {
-        try { await fs.promises.unlink(file.path) } catch (err) {
-          // ...
-        }
-      }
-    })
-  }
-
-  next()
+  Promise.all(tasks).then(x => next()).catch(x => next(x))
 }
 
 module.exports = [multerInstance, postProcess]
